orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Orchestrator agent for the financial advisory system."
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "google-cloud-secret-manager",
    "redis",
    "httpx",
    "google-generativeai",
    "structlog",
]
